Update prev aliases and wordlist env vars
Updated the nmapprev command due to issues with it working with the dependencies nmaptocsv. Removed it from usage and switched to awk. Updated the wordlists env vars because they were pointing at a non-existing directory. Removed nmaptocsv from the pipx install. Updated the ffufprev command to work with csvtomd.

Functionally nmapprev works the same way but now with colors.