reformat ffuf aliases
removed 302 from ffuf filtering list and updated the ffufprev alias to search for all .ffuf files instead. This was done in order to remove some of the overfiltering being done on results and input files. The other defined ffuf aliases needed to have their output files and usage texts updated.